requirements.txt
crewai
crewai-tools
streamlit 
python-dotenv
pandas
-e .